requirements.txt
# Abidjan Smart City Platform - Python Requirements

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2

# Kafka
kafka-python==2.0.2
confluent-kafka==2.3.0

# Databases
psycopg2-binary==2.9.9
pymongo==4.6.1
redis==5.0.1
SQLAlchemy==2.0.23
cassandra-driver==3.29.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.2
tensorflow==2.15.0
keras==2.15.0

# PySpark (installed separately or via conda)
# pyspark==3.4.1

# Security
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
httpx==0.25.2
locust==2.20.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
